Move interface to the correct package.
